:recycle: add base class for Candle management commands
