+++ quant_candles/management/base.py
@@ -2,10 +2,11 @@ import logging
 from typing import Optional
 
 from django.core.management.base import BaseCommand, CommandParser
+from django.db.models import QuerySet
 
 from quant_candles.constants import Exchange
 from quant_candles.lib import parse_period_from_to
-from quant_candles.models import Symbol
+from quant_candles.models import Candle, Symbol
 
 logger = logging.getLogger(__name__)
 
@@ -63,6 +64,45 @@ class BaseTradeDataCommand(BaseCommand):
             )
             return {
                 "symbol": symbol,
+                "timestamp_from": timestamp_from,
+                "timestamp_to": timestamp_to,
+            }
+
+
+class BaseCandleCommand(BaseCommand):
+    def get_queryset(self) -> QuerySet:
+        """Get queryset."""
+        raise NotImplementedError
+
+    def add_arguments(self, parser: CommandParser) -> None:
+        super().add_arguments(parser)
+        queryset = self.get_queryset()
+        parser.add_argument(
+            "name",
+            type=str,
+            choices=queryset.values_list("code_name", flat=True),
+        )
+        parser.add_argument("--date-to", type=str, default=None)
+        parser.add_argument("--time-to", type=str, default=None)
+        parser.add_argument("--date-from", type=str, default=None)
+        parser.add_argument("--time-from", type=str, default=None)
+        parser.add_argument("--retry", action="store_true")
+
+    def handle(self, *args, **options) -> None:
+        name = options["name"]
+        try:
+            candle = Candle.objects.get(code_name=name)
+        except Candle.DoesNotExist:
+            logger.warn(f"{name} not registered")
+        else:
+            timestamp_from, timestamp_to = parse_period_from_to(
+                date_from=options["date_from"],
+                time_from=options["time_from"],
+                date_to=options["date_to"],
+                time_to=options["time_to"],
+            )
+            return {
+                "candle": candle,
                 "timestamp_from": timestamp_from,
                 "timestamp_to": timestamp_to,
             }
